- re-organized existing tests
